Hurst Documentation

Hurst Documentation

Hurst Documentation

Index

Hurst.Simulation

Hurst.Calibration

Hurst.Visualisations

Hurst.Units

Hurst.Verification

Index

Hurst.Calibration.calibrate

Hurst.Simulation.simulate

Hurst.Units.acres_to_km2

Hurst.Units.cumecs_to_megalitres_day

Hurst.Units.km2_to_acres

Hurst.Units.km2_to_m2

Hurst.Units.m2_to_km2

Hurst.Units.megalitres_day_to_cumecs

Hurst.Units.megalitres_to_mm_runoff

Hurst.Units.mm_runoff_to_megalitres

Hurst.Verification.coeff_det

Hurst.Verification.kge

Hurst.Verification.kge

Hurst.Verification.mae

Hurst.Verification.mse

Hurst.Verification.nse

Hurst.Verification.persistence

Hurst.Verification.rmse

Hurst.Visualisations.hydrograph

Hurst.Visualisations.hyscatter

Hurst.Simulation

Hurst.Simulation.simulate — Method.

simulate(timestep_fnc, rain, pet, pars, init_state)

Returns a runoff simulation by forcing a rainfall-runoff model with the rain and pet timeseries.

The rainfall-runoff model is defined by the timestep_fnc function and pars model parameters Dictionary. This function is expected to return runoff for a single timestep of rain and pet. The model will be initalised with the init_state initial state Dictionary.

Will error if the rain and pet arrays are different lengths.

source

Hurst.Calibration

Hurst.Calibration.calibrate — Method.

calibrate(rain, pet, runoff, functions, opt_options)

Attempts to minimise an objective function between the runoff and a simulation from a rainfall-runoff model forced by rain and pet.

Returns an optimal set of parameters and the associated objective function value.

Will error if the rain, pet and runoff arrays are different lengths.

Functions to define the model and parameters are defined in the functions Dictionary. Refer to example to understand meaning.

functions[:run_model_time_step]
functions[:init_state]
functions[:params_from_array]
functions[:objective_function]
functions[:params_inverse_transform]
functions[:params_range_transform]
functions[:params_range_to_tuples]
functions[:params_range]

If the :params_inverse_transform key is defined in the functions Dictionary then it will assume the optimisation is in transformed space and will attempt to transform and detransform paramter sets and ranges. Will error if ALL the required transformation-related Dictionary keys are not defined.

Currently uses the BlackBoxOptim.jl package for numerical optimisation and the following options may be specified in the opt_options Dictionary, detailed information can be found at BlackBoxOptim.jl options:

opt_options[:method]
opt_options[:max_iterations]
opt_options[:max_time]
opt_options[:trace_interval]
opt_options[:trace_mode])

A typical set of function and opt_options to run a 5 minute calibration using the GR4J model in transformed paramter space with the Nash Sutcliffe Efficiency objective function could be:

# build up dictionary of model functions needed for calibration
functions = Dict()
functions[:run_model_time_step] = gr4j_run_step
functions[:init_state] = gr4j_init_state
functions[:params_from_array] = gr4j_params_from_array
functions[:objective_function] = (obs, sim) -> -1 * nse(obs, sim)
functions[:params_inverse_transform] = gr4j_params_trans_inv
functions[:params_range_transform] = gr4j_params_range_trans
functions[:params_range_to_tuples] = gr4j_params_range_to_tuples
functions[:params_range] = gr4j_params_range

# build up dictionary of optimiser options needed for calibration
opt_options = Dict()
opt_options[:max_iterations] = false
opt_options[:max_time] = 5 * 60
opt_options[:trace_interval] = 15
opt_options[:trace_mode] = :verbose
opt_options[:method] = :adaptive_de_rand_1_bin_radiuslimited

# calibrate the model
opt_pars, opt_neg_nse = calibrate(data[:rain], data[:pet], data[:runoff_obs], functions, opt_options)
opt_nse = -1 * opt_neg_nse

See also: simulate(timestep_fnc, rain, pet, pars, init_state)

source

Hurst.Visualisations

Hurst.Visualisations.hydrograph — Method.

hydrograph(rain, runoffs, runoff_labels)

Generates a simple combined hydrograph and hyetograph figure.

Multiple traces of runoff may be plotted by passing an array of arrays for runoff and associated runoff_labels.

Datetime tick marks are not plotted, only timesteps.

source

Hurst.Visualisations.hyscatter — Method.

hyscatter(obs, sim, labels)

Generates a simple scatter plot of a simulation series against an observation series.

Include a plot title, series label, and measurement units for the axis.

source

Hurst.Units

Hurst.Units.acres_to_km2 — Method.

acres_to_km2(a)

Convert an area in acres to an area in square kilometres.

See also: km2_to_acres(a)

source

Hurst.Units.cumecs_to_megalitres_day — Method.

cumecs_to_megalitres_day(q)

Convert cubic metres per second of streamflow q into megalitres per day of volume.

See also: megalitres_day_to_cumecs(v)

source

Hurst.Units.km2_to_acres — Method.

km2_to_acres(a)

Convert an area in square kilometres to an area in acres.

See also: acres_to_km2(a)

source

Hurst.Units.km2_to_m2 — Method.

km2_to_m2(a)

Convert an area in square kilometres to an area in square metres.

See also: m2_to_km2(a)

source

Hurst.Units.m2_to_km2 — Method.

m2_to_km2(a)

Convert an area in square metres to an area in square kilometres.

See also: km2_to_m2(a)

source

Hurst.Units.megalitres_day_to_cumecs — Method.

megalitres_day_to_cumecs(v)

Convert megalitres per day of volume v into cubic metres per second of streamflow.

See also: cumecs_to_megalitres_day(q)

source

Hurst.Units.megalitres_to_mm_runoff — Method.

megalitres_to_mm_runoff(v, area)

Use the catchment area (kilometres) to convert a volume v (megalitres) into catchment average runoff (millimetres).

See also: mm_runoff_to_megalitres(d, area)

source

Hurst.Units.mm_runoff_to_megalitres — Method.

mm_runoff_to_megalitres(d, area)

Use the catchment area (kilometres) to convert a depth d of catchment average runoff (millimetres) into volume (megalitres).

See also: megalitres_to_mm_runoff(v, area)

source

Hurst.Verification

Hurst.Verification.coeff_det — Method.

coeff_det(y, f)

Returns the Coefficient of Determination between y and f. Skips missing values from either series.

source

Hurst.Verification.kge — Method.

kge(o, s, components)

Returns the Kling-Gupta Efficiency between o and s. Skips missing values from either series.

If components is true then a Dictionary is returned containing the final KGE value (:kge) along with the individual components used to construct the KGE (:covariance, :relativevariability, :meanbias)

See also: kge(o, s)

source

Hurst.Verification.kge — Method.

kge(o, s)

Returns the Kling-Gupta Efficiency between o and s. Skips missing values from either series.

See also: kge(o, s, components)

source

Hurst.Verification.mae — Method.

mae(o, s)

Returns the Mean Absolute Error between o and s. Skips missing values from either series.

source

Hurst.Verification.mse — Method.

mse(o, s)

Returns the Mean Square Error between o and s. Skips missing values from either series.

See also: rmse(o, s)

source

Hurst.Verification.nse — Method.

nse(obs, sim)

Returns the Nash Sutcliffe Efficiency of obs and sim timeseries.

See also: coeff_det(y, f)

source

Hurst.Verification.persistence — Method.

persistence(o, s)

Returns the Persistence Index between o and s. Skips missing values from either series.

source

Hurst.Verification.rmse — Method.

rmse(o, s)

Returns the Root Mean Square Error between o and s.

See also: mse(o, s)

source