Hurst Documentation

Hurst Documentation

Hurst Documentation

Index

Hurst.Models.GR4J

Hurst.Models.OSTP

Hurst.Simulation

Hurst.Calibration

Hurst.Transformations

Hurst.Visualisations

Hurst.Units

Hurst.Performance.Metrics

Hurst.Performance.Economic

Hurst.Performance.Confusion

Index

Hurst.Calibration.calibrate

Hurst.Models.GR4J.create_uh_ordinates

Hurst.Models.GR4J.gr4j_init_state

Hurst.Models.GR4J.gr4j_param_trans

Hurst.Models.GR4J.gr4j_param_trans_inv

Hurst.Models.GR4J.gr4j_params_default

Hurst.Models.GR4J.gr4j_params_from_array

Hurst.Models.GR4J.gr4j_params_random

Hurst.Models.GR4J.gr4j_params_range

Hurst.Models.GR4J.gr4j_params_range_to_tuples

Hurst.Models.GR4J.gr4j_params_range_trans

Hurst.Models.GR4J.gr4j_params_to_array

Hurst.Models.GR4J.gr4j_params_trans

Hurst.Models.GR4J.gr4j_params_trans_inv

Hurst.Models.GR4J.gr4j_run_step

Hurst.Models.GR4J.s_curve

Hurst.Models.GR4J.update_uh

Hurst.Models.OSTP.ostp_init_state

Hurst.Models.OSTP.ostp_params_default

Hurst.Models.OSTP.ostp_params_from_array

Hurst.Models.OSTP.ostp_params_random

Hurst.Models.OSTP.ostp_params_range

Hurst.Models.OSTP.ostp_params_range_to_tuples

Hurst.Models.OSTP.ostp_params_to_array

Hurst.Models.OSTP.ostp_run_step

Hurst.Performance.Confusion.confusion

Hurst.Performance.Confusion.confusion_scaled

Hurst.Performance.Economic.cost_loss

Hurst.Performance.Economic.cost_loss_richardson

Hurst.Performance.Economic.cost_loss_roulin

Hurst.Performance.Economic.cost_loss_verkade

Hurst.Performance.Metrics.coeff_det

Hurst.Performance.Metrics.kge

Hurst.Performance.Metrics.kuipers_score

Hurst.Performance.Metrics.mae

Hurst.Performance.Metrics.mse

Hurst.Performance.Metrics.nse

Hurst.Performance.Metrics.persistence

Hurst.Performance.Metrics.rmse

Hurst.Simulation.simulate

Hurst.Transformations.boxcox

Hurst.Transformations.boxcox

Hurst.Transformations.boxcox_inverse

Hurst.Transformations.boxcox_inverse

Hurst.Transformations.log_sinh

Hurst.Transformations.log_sinh_inverse

Hurst.Transformations.log_trans

Hurst.Transformations.log_trans

Hurst.Transformations.log_trans_inverse

Hurst.Transformations.log_trans_inverse

Hurst.Units.acres_to_km2

Hurst.Units.cumecs_to_megalitres_day

Hurst.Units.km2_to_acres

Hurst.Units.km2_to_m2

Hurst.Units.m2_to_km2

Hurst.Units.megalitres_day_to_cumecs

Hurst.Units.megalitres_to_mm_runoff

Hurst.Units.mm_runoff_to_megalitres

Hurst.Visualisations.hydrograph

Hurst.Visualisations.hyscatter

Hurst.Models.GR4J

Hurst.Models.GR4J.gr4j_init_state — Method.

gr4j_init_state(pars)

Return a Dictionary containing an initial state for GR4J model. Uses a standard method derived from a set of model parmaters, pars.

The pars argument should have keys defined by the gr4j_params_from_array function.

Essential input for the first call to gr4j_run_step function to ensure the unit hydrograph arrays are initialised to the correct size and ordinates are specified correctly.

See also: gr4j_params_from_array, gr4j_run_step

source

Hurst.Models.GR4J.gr4j_params_default — Method.

gr4j_params_default()

Returns a Dictionary containing a reasonable set of GR4J parameter vaues with keys defined by the gr4j_params_from_array function and ready to be used by the gr4j_run_step function.

See also: gr4j_params_from_array, gr4j_run_step

source

Hurst.Models.GR4J.gr4j_params_from_array — Method.

gr4j_params_from_array(arr)

Returns a Dictionary containing a parameter set ready to be used by the gr4j_run_step function.

Sets the value of the :x1, :x2, :x3 and :x4 keys to the 1st, 2nd, 3rd and 4th index values of the arr array.

Intentionally does not check if these parameter values are within any acceptable range.

See also: gr4j_params_to_array, gr4j_params_range, gr4j_run_step

source

Hurst.Models.GR4J.gr4j_params_random — Method.

gr4j_params_random(prange)

Returns a Dictionary containing a random set of GR4J parameter vaues selected with a uniform sampler within the parameter ranges specified by prange.

The prange argument should be a Dictionary with keys defined by the gr4j_params_range function.

This set of random parameters has keys defined by the gr4j_params_from_array function and is ready to be used by the gr4j_run_step function.

See also: gr4j_params_from_array, gr4j_run_step, gr4j_params_range

source

Hurst.Models.GR4J.gr4j_params_range — Method.

gr4j_params_range()

Returns a Dictionary with reasonable ranges for GR4J parameter vaues.

These are used in the model calibration and random parameter sampling functions.

See also: gr4j_params_from_array, gr4j_run_step, gr4j_params_random

source

Hurst.Models.GR4J.gr4j_params_range_to_tuples — Method.

gr4j_params_range_to_tuples(prange)

Returns an array of tuples containing the GR4J parameter ranges provided in the prange Dictionary argument. This Dictionary should contain the keys defined by the gr4j_params_range function.

See also: gr4j_params_range

source

Hurst.Models.GR4J.gr4j_params_range_trans — Method.

gr4j_params_range_trans(prange)

Returns a Dictionary with reasonable ranges for GR4J parameter vaues which have been transformed using the method defined in the gr4j_param_trans function.

See also: gr4j_params_range, gr4j_run_step, gr4j_params_random

source

Hurst.Models.GR4J.gr4j_params_to_array — Method.

gr4j_params_to_array(pars)

Returns an array containing GR4J parameter vaues from a parameter Dictionary with keys defined by the gr4j_params_from_array function.

Sets the 1st, 2nd, 3rd and 4th index values of the returned array to the value of the :x1, :x2, :x3 and :x4 keys in the pars Dictionary.

Intentionally does not check if these parameter values are within any acceptable range.

See also: gr4j_params_from_array, gr4j_params_range

source

Hurst.Models.GR4J.gr4j_params_trans — Method.

gr4j_params_trans(pars)

Returns a set of GR4J parameters which has been transformed using standard practice for a more uniform parameter search space to calibrate within.

The pars Dictionary should correspond to the that generated by the gr4j_params_from_array function.

See also: gr4j_params_trans_inv, gr4j_params_from_array

source

Hurst.Models.GR4J.gr4j_params_trans_inv — Method.

gr4j_params_trans_inv(pars)

Returns a set of transformed GR4J parameters pars which have been back-transformed. The back-transformation performed is the inverse of the transformations used in the function gr4j_params_trans.

The pars Dictionary should correspond to the that generated by the gr4j_params_from_array function.

See also: gr4j_params_trans, gr4j_params_from_array

source

Hurst.Models.GR4J.gr4j_run_step — Method.

gr4j_run_step(rain, pet, state, pars)

Run a single time-step of the GR4J model. Model is forced by the rain and pet (floats) supplied and uses the state for initial conditions. Model parameters used are provided in the pars argument.

The pars argument should have keys defined by the gr4j_params_from_array function and the state with keys defined by gr4j_init_state.

The function then returns the runoff and an updated state. This updated state is typically used as the input state for the next time-step in a time-series simulation.

See also: gr4j_init_state, gr4j_params_from_array

source

Hurst.Models.GR4J.create_uh_ordinates — Method.

create_uh_ordinates(variant, size, x4)

Returns an Array of size elements containing the GR4J unit hydrograph ordinates for either of the two variant parameterised by the x4 parameter.

Uses the s_curve function to determine the ordinate values and is used by the gr4j_init_state function to create an initial state for GR4J, specifically defining the unit hydrograph ordinate values used when updating the unit hydrographs each time-step in the gr4j_run_step function.

See also: s_curve, gr4j_init_state, gr4j_run_step

source

Hurst.Models.GR4J.gr4j_param_trans — Method.

gr4j_param_trans(param, value)

Returns the value of a single GR4J parameter which has been transformed using standard practice for a more uniform parameter search space to calibrate within.

The param Symbol should correspond to a key in the Dictionary returned by the gr4j_params_from_array function.

See also: gr4j_param_trans_inv, gr4j_params_from_array, gr4j_params_trans

source

Hurst.Models.GR4J.gr4j_param_trans_inv — Method.

gr4j_param_trans_inv(param, value)

Returns a single transformed GR4J parameter value which has been back-transformed. The back-transformation performed is the inverse of the transformations defined in the function gr4j_param_trans.

The param Symbol should correspond to a key in the Dictionary returned by the gr4j_params_from_array function.

See also: [gr4j_param_trans(param, value), [gr4jparamsfromarray](@ref), [gr4jparam_trans`](@ref)

source

Hurst.Models.GR4J.s_curve — Method.

s_curve(variant, scale, x)

Returns the value at location x of a variant of an s-curve function parameterised with a scale.

Is used to define the values of the GR4J unit hydrograph ordinates in the function create_uh_ordinates.

See also: create_uh_ordinates, gr4j_run_step

source

Hurst.Models.GR4J.update_uh — Method.

update_uh(uh, volume, ordinates)

Returns an updated unit hydrograph Array after incrementing the uh and convoluting with the ordinates and volume.

The ordinates are defined by the create_uh_ordinates function.

Used by the gr4j_run_step function when updating the model state, specifically the two GR4J unit hydrographs.

See also: create_uh_ordinates, gr4j_run_step

source

Hurst.Models.OSTP

Hurst.Models.OSTP.ostp_init_state — Method.

ostp_init_state(pars)

Return a Dictionary containing an initial state for ostp model. Uses a standard method derived from a set of model parmaters, pars.

The pars argument should have keys defined by the ostp_params_from_array function.

See also: ostp_params_from_array, ostp_run_step

source

Hurst.Models.OSTP.ostp_params_default — Method.

ostp_params_default()

Returns a Dictionary containing a reasonable set of ostp parameter vaues with keys defined by the ostp_params_from_array function and ready to be used by the ostp_run_step function.

See also: ostp_params_from_array, ostp_run_step

source

Hurst.Models.OSTP.ostp_params_from_array — Method.

ostp_params_from_array(arr)

Returns a Dictionary containing a parameter set ready to be used by the ostp_run_step function.

Sets the value of the :capacity, :loss keys to the 1st and 2nd index values of the arr array.

Intentionally does not check if these parameter values are within any acceptable range.

See also: ostp_params_to_array, ostp_params_range, ostp_run_step

source

Hurst.Models.OSTP.ostp_params_random — Method.

ostp_params_random(prange)

Returns a Dictionary containing a random set of ostp parameter vaues selected with a uniform sampler within the parameter ranges specified by prange.

The prange argument should be a Dictionary with keys defined by the ostp_params_range function.

This set of random parameters has keys defined by the ostp_params_from_array function and is ready to be used by the ostp_run_step function.

See also: ostp_params_from_array, ostp_run_step, ostp_params_range

source

Hurst.Models.OSTP.ostp_params_range — Method.

ostp_params_range()

Returns a Dictionary with reasonable ranges for ostp parameter vaues.

These are used in the model calibration and random parameter sampling functions.

See also: ostp_params_from_array, ostp_run_step, ostp_params_random

source

Hurst.Models.OSTP.ostp_params_range_to_tuples — Method.

ostp_params_range_to_tuples(prange)

Returns an array of tuples containing the ostp parameter ranges provided in the prange Dictionary argument. This Dictionary should contain the keys defined by the ostp_params_range function.

See also: ostp_params_range

source

Hurst.Models.OSTP.ostp_params_to_array — Method.

ostp_params_to_array(pars)

Returns an array containing ostp parameter vaues from a parameter Dictionary with keys defined by the ostp_params_from_array function.

Sets the 1st and 2nd index values of the returned array to the value of the :capacity and :loss keys in the pars Dictionary.

Intentionally does not check if these parameter values are within any acceptable range.

See also: ostp_params_from_array, ostp_params_range

source

Hurst.Models.OSTP.ostp_run_step — Method.

ostp_run_step(rain, pet, state, pars)

Run a single time-step of the ostp model. ostp a simple one storage, two parameter model used for educational purposes. To outline basic concepts in conceptual rainfall-runoff models and how to implement a rainfall-runoff model within Hurst. Please do not use it for actual water resource modelling!

The model is forced by the rain and pet (floats) supplied and uses the state for initial conditions. Model parameters used are provided in the pars argument.

The pars argument should have keys defined by the ostp_params_from_array function and the state with keys defined by ostp_init_state.

The function then returns the runoff and an updated state. This updated state is typically used as the input state for the next time-step in a time-series simulation.

See also: ostp_init_state, ostp_params_from_array

source

Hurst.Simulation

Hurst.Simulation.simulate — Method.

simulate(timestep_fnc, rain, pet, pars, init_state)

Returns a runoff simulation by forcing a rainfall-runoff model with the rain and pet timeseries.

The rainfall-runoff model is defined by the timestep_fnc function and pars model parameters Dictionary. This function is expected to return runoff for a single timestep of rain and pet. The model will be initalised with the init_state initial state Dictionary.

Will error if the rain and pet arrays are different lengths.

source

Hurst.Calibration

Hurst.Calibration.calibrate — Method.

calibrate(rain, pet, runoff, functions, opt_options)

Attempts to minimise an objective function between the runoff and a simulation from a rainfall-runoff model forced by rain and pet.

Returns an optimal set of parameters and the associated objective function value.

Will error if the rain, pet and runoff arrays are different lengths.

Functions to define the model and parameters are defined in the functions Dictionary. Refer to example to understand meaning.

functions[:run_model_time_step]
functions[:init_state]
functions[:params_from_array]
functions[:objective_function]
functions[:params_inverse_transform]
functions[:params_range_transform]
functions[:params_range_to_tuples]
functions[:params_range]

If the :params_inverse_transform key is defined in the functions Dictionary then it will assume the optimisation is in transformed space and will attempt to transform and detransform paramter sets and ranges. Will error if ALL the required transformation-related Dictionary keys are not defined.

Currently uses the BlackBoxOptim.jl package for numerical optimisation and the following options may be specified in the opt_options Dictionary, detailed information can be found at BlackBoxOptim.jl options:

opt_options[:method]
opt_options[:max_iterations]
opt_options[:max_time]
opt_options[:trace_interval]
opt_options[:trace_mode])

A typical set of function and opt_options to run a 5 minute calibration using the GR4J model in transformed paramter space with the Nash Sutcliffe Efficiency objective function could be:

# build up dictionary of model functions needed for calibration
functions = Dict()
functions[:run_model_time_step] = gr4j_run_step
functions[:init_state] = gr4j_init_state
functions[:params_from_array] = gr4j_params_from_array
functions[:objective_function] = (obs, sim) -> -1 * nse(obs, sim)
functions[:params_inverse_transform] = gr4j_params_trans_inv
functions[:params_range_transform] = gr4j_params_range_trans
functions[:params_range_to_tuples] = gr4j_params_range_to_tuples
functions[:params_range] = gr4j_params_range

# build up dictionary of optimiser options needed for calibration
opt_options = Dict()
opt_options[:max_iterations] = false
opt_options[:max_time] = 5 * 60
opt_options[:trace_interval] = 15
opt_options[:trace_mode] = :verbose
opt_options[:method] = :adaptive_de_rand_1_bin_radiuslimited

# calibrate the model
opt_pars, opt_neg_nse = calibrate(data[:rain], data[:pet], data[:runoff_obs], functions, opt_options)
opt_nse = -1 * opt_neg_nse

See also: simulate

source

Hurst.Transformations

Hurst.Transformations.boxcox — Method.

boxcox(y, λ, ν)

Two param box-cox transform. (Wikipedia)

Box, G. E. P., and D. R. Cox (1964), An analysis of transformations, J. R. Stat. Soc., Ser. B, 26, 296–311.

See also: boxcox_inverse

source

Hurst.Transformations.boxcox — Method.

boxcox(y, λ)

One param box-cox transform. (Wikipedia)

Box, G. E. P., and D. R. Cox (1964), An analysis of transformations, J. R. Stat. Soc., Ser. B, 26, 296–311.

See also: boxcox_inverse

source

Hurst.Transformations.boxcox_inverse — Method.

boxcox_inverse(z, λ, ν)

Inverse of the two param box-cox transform. (Wikipedia)

Box, G. E. P., and D. R. Cox (1964), An analysis of transformations, J. R. Stat. Soc., Ser. B, 26, 296–311.

See also: boxcox

source

Hurst.Transformations.boxcox_inverse — Method.

boxcox_inverse(z, λ)

Inverse of the one param box-cox transform. (Wikipedia)

Box, G. E. P., and D. R. Cox (1964), An analysis of transformations, J. R. Stat. Soc., Ser. B, 26, 296–311.

See also: boxcox

source

Hurst.Transformations.log_sinh — Method.

log_sinh(y, a, b)

Log-sinh transform.

Wang, Q. J., D. L. Shrestha, D. E. Robertson, and P. Pokhrel (2012b), A log-sinh transformation for data normalization and variance stabiliza- tion, Water Resour. Res., 48, W05514, doi:10.1029/2011WR010973.

See also: log_sinh_inverse

source

Hurst.Transformations.log_sinh_inverse — Method.

log_sinh_inverse(z, a, b)

Inverse of the log-sinh transform.

Wang, Q. J., D. L. Shrestha, D. E. Robertson, and P. Pokhrel (2012b), A log-sinh transformation for data normalization and variance stabiliza- tion, Water Resour. Res., 48, W05514, doi:10.1029/2011WR010973.

See also: log_sinh

source

Hurst.Transformations.log_trans — Method.

log_trans(y, offset)

Log transform with offset. (Wikipedia)

See also: log_trans_inverse

source

Hurst.Transformations.log_trans — Method.

log_trans(y)

Log transform with no offset. (Wikipedia)

See also: log_trans_inverse

source

Hurst.Transformations.log_trans_inverse — Method.

log_trans_inverse(z, offset)

Inverse log transform with offset. (Wikipedia)

See also: log_trans

source

Hurst.Transformations.log_trans_inverse — Method.

log_trans_inverse(z)

Inverse log transform with no offset. (Wikipedia)

See also: log_trans

source

Hurst.Visualisations

Hurst.Visualisations.hydrograph — Method.

hydrograph(rain, runoffs, runoff_labels)

Generates a simple combined hydrograph and hyetograph figure.

Multiple traces of runoff may be plotted by passing an array of arrays for runoff and associated runoff_labels.

Datetime tick marks are not plotted, only timesteps.

source

Hurst.Visualisations.hyscatter — Method.

hyscatter(obs, sim, sim_labels)

Generates a simple scatter plot of simulation series sims against an observation series obs.

Multiple scatters may be plotted by passing an array of arrays for sims and associated sim_labels.

Include a plot title, series sim_labels, and measurement units for the axis.

source

Hurst.Units

Hurst.Units.acres_to_km2 — Method.

acres_to_km2(a)

Convert an area in acres to an area in square kilometres.

See also: km2_to_acres

source

Hurst.Units.cumecs_to_megalitres_day — Method.

cumecs_to_megalitres_day(q)

Convert cubic metres per second of streamflow q into megalitres per day of volume.

See also: megalitres_day_to_cumecs

source

Hurst.Units.km2_to_acres — Method.

km2_to_acres(a)

Convert an area in square kilometres to an area in acres.

See also: acres_to_km2

source

Hurst.Units.km2_to_m2 — Method.

km2_to_m2(a)

Convert an area in square kilometres to an area in square metres.

See also: m2_to_km2

source

Hurst.Units.m2_to_km2 — Method.

m2_to_km2(a)

Convert an area in square metres to an area in square kilometres.

See also: km2_to_m2

source

Hurst.Units.megalitres_day_to_cumecs — Method.

megalitres_day_to_cumecs(v)

Convert megalitres per day of volume v into cubic metres per second of streamflow.

See also: cumecs_to_megalitres_day

source

Hurst.Units.megalitres_to_mm_runoff — Method.

megalitres_to_mm_runoff(v, area)

Use the catchment area (kilometres) to convert a volume v (megalitres) into catchment average runoff (millimetres).

See also: mm_runoff_to_megalitres

source

Hurst.Units.mm_runoff_to_megalitres — Method.

mm_runoff_to_megalitres(d, area)

Use the catchment area (kilometres) to convert a depth d of catchment average runoff (millimetres) into volume (megalitres).

See also: megalitres_to_mm_runoff

source

Hurst.Performance.Metrics

Hurst.Performance.Metrics.coeff_det — Method.

coeff_det(y, f)

Returns the Coefficient of Determination between y and f. Skips missing values from either series.

source

Hurst.Performance.Metrics.kge — Method.

kge(o, s, return_components)

Returns the Kling-Gupta Efficiency between o and s. Skips missing values from either series.

If return_components is true then a Dictionary is returned containing the final KGE value (:kge) along with the individual returncomponents used to construct the KGE (:covariance, :relativevariability, :mean_bias)

source

Hurst.Performance.Metrics.kuipers_score — Method.

kuipers_score(scaled_conf)

Returns the Kuipers Score using the elements of the scaled_conf contingency table.

Richardson, D. S. “Skill and Relative Economic Value of the ECMWF Ensemble Prediction System.” Quarterly Journal of the Royal Meteorological Society 126, no. 563 (January 2000): 649–67. https://doi.org/10.1256/smsqj.56312.

See also: confusion_scaled

source

Hurst.Performance.Metrics.mae — Method.

mae(o, s)

Returns the Mean Absolute Error between o and s. Skips missing values from either series.

source

Hurst.Performance.Metrics.mse — Method.

mse(o, s)

Returns the Mean Square Error between o and s. Skips missing values from either series.

See also: rmse

source

Hurst.Performance.Metrics.nse — Method.

nse(obs, sim)

Returns the Nash Sutcliffe Efficiency of obs and sim timeseries.

See also: coeff_det

source

Hurst.Performance.Metrics.persistence — Method.

persistence(o, s)

Returns the Persistence Index between o and s. Skips missing values from either series.

source

Hurst.Performance.Metrics.rmse — Method.

rmse(o, s)

Returns the Root Mean Square Error between o and s.

See also: mse

source

Hurst.Performance.Economic

Hurst.Performance.Economic.cost_loss — Method.

cost_loss(costs, losses, scaled_conf, method="verkade")

Returns the relative economic value of a forecast system using a the cost-loss model using the cost-loss ratio (costs, losses) and scaled confusion matrix scaled_conf.

Two methods are implemented (Verkade 2011 and Roulin 2007) which may be selected with the method argument (default is Verkade).

Verkade, J. S., and M. G. F. Werner. “Estimating the Benefits of Single Value and Probability Forecasting for Flood Warning.” Hydrology and Earth System Sciences 15, no. 12 (December 20, 2011): 3751–65. https://doi.org/10.5194/hess-15-3751-2011.

Roulin, E. “Skill and Relative Economic Value of Medium-Range Hydrological Ensemble Predictions.” Hydrology and Earth System Sciences 11, no. 2 (2007): 725–37. https://doi.org/10.5194/hess-11-725-2007.

See also: confusion_scaled, cost_loss_roulin, cost_loss_verkade, cost_loss_richardson

source

Hurst.Performance.Economic.cost_loss_richardson — Method.

cost_loss_richardson(α, μ, H, F)

Returns the relative economic value of a forecast system using a the cost-loss model using the cost-loss ratio α, observed frequency μ, hit rate H and false alarm rate F.

This method has the same implementation as Roulin 2007 but does not use a confusion matrix as input.

Richardson, D. S. “Skill and Relative Economic Value of the ECMWF Ensemble Prediction System.” Quarterly Journal of the Royal Meteorological Society 126, no. 563 (January 2000): 649–67. https://doi.org/10.1256/smsqj.56312.

See also: confusion_scaled, cost_loss, cost_loss_verkade, cost_loss_roulin

source

Hurst.Performance.Economic.cost_loss_roulin — Method.

cost_loss_roulin(costs, losses, scaled_conf)

Returns the relative economic value of a forecast system using a the cost-loss model using the cost-loss ratio (costs, losses) and confusion matrix scaled by the total events and non-events scaled_conf.

Roulin, E. “Skill and Relative Economic Value of Medium-Range Hydrological Ensemble Predictions.” Hydrology and Earth System Sciences 11, no. 2 (2007): 725–37. https://doi.org/10.5194/hess-11-725-2007.

See also: confusion_scaled, cost_loss, cost_loss_verkade, cost_loss_richardson

source

Hurst.Performance.Economic.cost_loss_verkade — Method.

cost_loss_verkade(costs, losses, scaled_conf)

Returns the relative economic value of a forecast system using a the cost-loss model using the cost-loss ratio (costs, losses) and confusion matrix scaled by the total events and non-events scaled_conf.

Verkade, J. S., and M. G. F. Werner. “Estimating the Benefits of Single Value and Probability Forecasting for Flood Warning.” Hydrology and Earth System Sciences 15, no. 12 (December 20, 2011): 3751–65. https://doi.org/10.5194/hess-15-3751-2011.

See also: confusion_scaled, cost_loss, cost_loss_roulin, cost_loss_richardson

source

Hurst.Performance.Confusion

Hurst.Performance.Confusion.confusion — Method.

confusion(hits, misses, false_alarms, quiets)

Returns a Dict containing fields of a confusion matrix with the various common names used for each combination of the 2x2 grid.

See also: confusion_scaled

source

Hurst.Performance.Confusion.confusion_scaled — Method.

confusion_scaled(hits, misses, false_alarms, quiets)

Returns a Dict containing fields of a confusion matrix with the various common names used for each combination of the 2x2 grid scaled to be relative to the number of events and non-events (hits + misses + false_alarms + quiets).

See also: confusion

source